+++ string/tests/sa_nadc21h.cpp
@@ -2,8 +2,8 @@
 using namespace std;
 
 #ifdef HENRYX
-#include "../../lca/string/suffix_array.h"
-#include "../../lca/data_structure/min_queue.h"
+#include "../suffix_array.h"
+#include "../../data_structure/min_queue.h"
 #else
 #include "suffix_array.h"
 #include "min_queue.h"
@@ -21,7 +21,7 @@ int main() {
   for(int i=0; i<n; i++) {
     string s;
     cin >> s;
-    len.push_back(size(s) + 1);
+    len.push_back((int)size(s) + 1);
     cat.push_back('z' + (int)size(len));
     cat.insert(end(cat), begin(s), end(s));
   }
@@ -36,14 +36,14 @@ int main() {
   map<int,int> have;
   min_queue<int> window;
   int ans = 0;
-  for(int i=1, j=1; i<=size(cat); i++) {
-    while(j <= size(cat) && size(have) < n) {
+  for(int i=1, j=1; i<=(int)size(cat); i++) {
+    while(j <= (int)size(cat) && (int)size(have) < n) {
       if(j > i) window.push(sa.height[j]);
-      have[upper_bound(begin(len), end(len), sa[j++]) - begin(len)] += 1;
+      have[int(upper_bound(begin(len), end(len), sa[j++]) - begin(len))] += 1;
     }
-    if(size(have) < n) break;
+    if((int)size(have) < n) break;
     ans = max(ans, window.min());
-    int rem = upper_bound(begin(len), end(len), sa[i]) - begin(len);
+    int rem = int(upper_bound(begin(len), end(len), sa[i]) - begin(len));
     if(--have[rem] == 0) have.erase(rem);
     window.pop();
   }
