+++ string/tests/sa_spoj_LCS2.cpp
@@ -12,12 +12,12 @@ int main() {
   basic_string<int> cat;
   vector<int> len;
   for(string s; cin >> s; ) {
-    len.push_back(size(s) + 1);
+    len.push_back((int)size(s) + 1);
     cat.push_back('z' + (int)size(len));
     cat.insert(end(cat), begin(s), end(s));
   }
 
-  int n = size(len);
+  int n = (int)size(len);
   if (n == 1) {
     cout << len[0] - 1 << nl;
     return 0;
@@ -28,14 +28,14 @@ int main() {
   map<int,int> have;
   min_queue<int> window;
   int ans = 0;
-  for(int i=1, j=1; i<=size(cat); i++) {
-    while(j <= size(cat) && size(have) < n) {
+  for(int i=1, j=1; i<=(int)size(cat); i++) {
+    while(j <= (int)size(cat) && (int)size(have) < n) {
       if(j > i) window.push(sa.height[j]);
-      have[upper_bound(begin(len), end(len), sa[j++]) - begin(len)] += 1;
+      have[int(upper_bound(begin(len), end(len), sa[j++]) - begin(len))] += 1;
     }
-    if(size(have) < n) break;
+    if((int)size(have) < n) break;
     ans = max(ans, window.min());
-    int rem = upper_bound(begin(len), end(len), sa[i]) - begin(len);
+    int rem = int(upper_bound(begin(len), end(len), sa[i]) - begin(len));
     if(--have[rem] == 0) have.erase(rem);
     window.pop();
   }
